person类

+++ src/main/java/hello/Repository/PersonRepository.java
@@ -1,9 +1,9 @@
 package hello.Repository;
 
 import hello.Entity.Person;
+
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Long> {
-
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 
 }
